Exception handling - delete-Ajuste erro 204

+++ src/main/java/com/pessoal/webservices/services/UserService.java
@@ -15,38 +15,41 @@ import com.pessoal.webservices.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	private UserRepository userRepository;
-	
-	public List<User> findAll(){
-		
+
+	public List<User> findAll() {
+
 		return userRepository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = userRepository.findById(id);
-		
+
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return userRepository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
 		try {
-		userRepository.deleteById(id);
-		
-		}
-		catch(EmptyResultDataAccessException e) {
+			if (!userRepository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			}
+
+			userRepository.deleteById(id);
+		} 
+		catch (EmptyResultDataAccessException e) {
 			throw new ResourceNotFoundException(id);
-		}
-		catch(DataIntegrityViolationException e) {
+		} 
+		catch (DataIntegrityViolationException e) {
 			throw new DatabaseException(e.getMessage());
 		}
 	}
-	
+
 	public User update(Long id, User obj) {
 		User user = userRepository.getReferenceById(id);
 		upDateData(user, obj);
